Replace asm with RtlCaptureContext
We lose saving of fp registers, but meh
